put tests back into index.test.js
